compiles to the same file extension
I was having hard time on compiling files to more than one extension. For example, I have three files, one with .php extension (having html + php code in it) and other two with .html extension. When I compiled these files, all three files were compiling with extension .html. So instead, I made modifications which will let the files to compile as it is with the user provided extension when the user hadn't mention the "ext" option.